Finished adding Weather API.

+++ src/Tap.Dotnet.Web.Application/Models/WeatherForecastViewModel.cs
@@ -4,9 +4,9 @@
     {
         public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+        public Single TemperatureC { get; set; }
 
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+        public Single TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
         public string? Description { get; set; }
     }
